junit fixes set font size via Double color handling via HSL calculation
git-svn-id: https://svn.apache.org/repos/asf/poi/branches/common_sl@1691117 13f79535-47bb-0310-9956-ffa450edef68

+++ src/scratchpad/src/org/apache/poi/hslf/model/PPGraphics2D.java
@@ -256,7 +256,7 @@ public final class PPGraphics2D extends Graphics2D implements Cloneable {
         txt.setText(s);
 
         HSLFTextRun rt = txt.getTextParagraphs().get(0).getTextRuns().get(0);
-        rt.setFontSize(_font.getSize());
+        rt.setFontSize((double)_font.getSize());
         rt.setFontFamily(_font.getFamily());
 
         if (getColor() != null) rt.setFontColor(getColor());
